ft:openai api app for generating task and autograding
