+++ package.json
@@ -14,39 +14,40 @@
     "start": "next start"
   },
   "dependencies": {
-    "@ai-sdk/anthropic": "^0.0.48",
+    "@ai-sdk/openai": "^0.0.60",
     "@t3-oss/env-nextjs": "^0.10.1",
     "@tanstack/react-query": "^5.54.1",
     "@types/uuid": "^10.0.0",
-    "ai": "^3.3.24",
+    "ai": "^3.3.39",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
     "drizzle-orm": "^0.33.0",
-    "geist": "^1.3.0",
-    "next": "^14.2.4",
+    "geist": "^1.3.1",
+    "next": "^14.2.7",
     "postgres": "^3.4.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
+    "react-toastify": "^10.0.5",
     "uuid": "^10.0.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/eslint": "^8.56.10",
-    "@types/node": "^20.14.10",
-    "@types/react": "^18.3.3",
+    "@types/eslint": "^8.56.12",
+    "@types/node": "^20.16.3",
+    "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^8.1.0",
-    "@typescript-eslint/parser": "^8.1.0",
-    "drizzle-kit": "^0.24.0",
+    "@typescript-eslint/eslint-plugin": "^8.3.0",
+    "@typescript-eslint/parser": "^8.3.0",
+    "drizzle-kit": "^0.24.2",
     "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.4",
+    "eslint-config-next": "^14.2.7",
     "eslint-plugin-drizzle": "^0.2.3",
-    "postcss": "^8.4.39",
-    "prettier": "^3.3.2",
-    "prettier-plugin-tailwindcss": "^0.6.5",
-    "tailwindcss": "^3.4.3",
-    "typescript": "^5.5.3"
+    "postcss": "^8.4.44",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.6",
+    "tailwindcss": "^3.4.10",
+    "typescript": "^5.5.4"
   },
   "ct3aMetadata": {
     "initVersion": "7.37.0"
