+++ src/utils/validation/answersSchema.validation.ts
@@ -1,10 +1,9 @@
 import { z } from 'zod';
-import { v4 as uuidv4 } from 'uuid';
 
 export const AnswerSchema = z.object({
-  id: z.string().uuid().default(() => uuidv4()),
-  questionId: z.string(),
+  questionId:z.string(),
   userAnswers: z.array(z.string()),
+  expectedAnswers: z.array(z.string()),
   grading: z.number().min(0).max(5),
 });
